Duplication and atlas update (#466)
* Added Duplicate function

Added to the nodes right click menu. There may be some variables that were missed, particularly animations.

* Update nodes.d

* Update nodes.d

* Atlas update

Atlassing switched to per texture instead of based on UUid

* User friendliness

Updated the duplication to place the new object next in the order of children

* Cleaned the code

Sorry about that, To be totally honest, I wrote that before I even discovered not was !is. !cast was a game changer