Added compatible pairing copy
Would prefer to remove the origonal paste as i don't really see its usecase, so an ugly workaround to use axis 2 was added for a different function.

This version is nessisary to get a clean mirror on a negative scaled object, or any copied object. The origional paste or mirror will always cause imperfections.